# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Kestrel AI backend API
  version: 1.0.0
  description: |
    Contract between the Tauri/React frontend and the FastAPI sidecar. This file is the
    source of truth (spec section 2). The backend validates responses against it in tests,
    the TypeScript client is generated from it, and Prism serves it as the mock server.

    Conventions:
    - Every path below carries the `/api/v1` prefix. JSON bodies, UTF-8.
    - Every request carries the per-launch token, either as `Authorization: Bearer <token>`
      or as a `token` query parameter (needed for `<img>` sources and websocket connects).
    - Ids are UUID strings. Timestamps are RFC 3339 strings in UTC.
    - Boxes are in image pixels: `x, y` top-left corner, `w, h` size.
    - Every list endpoint paginates with `limit` and `cursor` and returns `next_cursor`
      (`null` when there is no further page).
    - Every error uses the `Error` envelope `{error: {code, message, details}}`.
    - Long work returns a `Job`; progress arrives over the websocket described in
      `x-websocket` and can be polled through the `jobs` resource.
servers:
  - url: http://127.0.0.1:8765
    description: development backend origin (the packaged app uses a random free port)
  - url: http://127.0.0.1:4010
    description: Prism mock server
security:
  - bearerAuth: []
  - tokenQuery: []
...
paths:
...
  /api/v1/projects:
    get:
      tags: [projects]
      operationId: listProjects
      summary: Recently opened projects, most recent first. Folders that no longer exist are skipped.
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: recent projects
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectPage" }
        default: { $ref: "#/components/responses/Error" }
    post:
      tags: [projects]
      operationId: createProject
      summary: Create a project in an empty or new folder and open it.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectCreate" }
      responses:
        "201":
          description: created and opened
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "409":
          description: the folder already contains a project (`code` is `already_exists`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/open:
    post:
      tags: [projects]
      operationId: openProject
      summary: Open an existing project folder (one that contains `project.db`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectOpen" }
      responses:
        "200":
          description: opened
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "404":
          description: no project in that folder
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [projects]
      operationId: getProject
      responses:
        "200":
          description: the project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      tags: [projects]
      operationId: updateProject
      summary: Update name, pre-annotation model or import defaults (project settings screen).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdate" }
      responses:
        "200":
          description: updated project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [projects]
      operationId: forgetProject
      summary: Remove the project from the recent list. The project folder and everything in it stay on disk; "Open folder" brings it back.
      responses:
        "204":
          description: forgotten
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/classes:
    parameters:
      - $ref: "#/components/parameters/projectId"
    put:
      tags: [projects]
      operationId: updateClasses
      summary: |
        Replace the class list. Items with an `id` keep it (rename, recolour, rehotkey);
        items without one are new; order is the list order. A class that still has boxes
        cannot be removed: the request fails with 409 `class_in_use` and the caller must
        reassign or delete the boxes first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/ClassDefInput" }
      responses:
        "200":
          description: updated project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "409":
          description: "a removed class still has boxes (`code` is `class_in_use`, details `{class_id, box_count}`), or two classes share a name or a hotkey, or a name is blank (`code` is `conflict`)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/stats:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [projects]
      operationId: getProjectStats
      summary: Statistics over all sources (spec section 5).
      responses:
        "200":
          description: statistics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stats" }
        default: { $ref: "#/components/responses/Error" }

  # ------------------------------------------------------------------ sources
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: the per-launch token passed to the sidecar in `APP_TOKEN`
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: same token as a query parameter, for image URLs and the websocket
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
    cursor:
      name: cursor
      in: query
      description: opaque cursor from the previous page's `next_cursor`
      schema: { type: string }
...
  responses:
    Error:
      description: error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              description: |
                machine-readable: unauthorized, not_found, validation_error, already_exists,
                class_in_use, conflict, not_implemented, provider_error, internal_error,
                wrong_project_kind (409: the operation does not belong to this kind of project;
                details `{kind, allowed}`), library_unavailable (503: the model library could not
                be opened), model_unavailable (409: the library model's weights file is missing),
                unmapped_classes (422: the run's model has classes with no project class and no
                remembered mapping; details `{model_id, unmapped}`), unsupported_point_cloud,
                insufficient_memory and insufficient_disk (422: a point cloud cannot be read or
                admitted), link_needs_coordinates, no_overlap and crs_already_set (422: a point
                cloud patch), invalid_epsg (422: an EPSG code pyproj does not know),
                wrong_point_count, needs_projected_crs, vertical_span_too_small and
                measurement_limit (422: a point-cloud measurement), octree_missing (404: a ready
                cloud's display-copy file is missing on disk), grid_too_large (422: a surface
                grid over the cell ceiling), unsupported_crs, source_missing and
                invalid_build_request (422: a surface build), no_coordinates (422: an ortho
                underlay without a CRS), invalid_geometry and invalid_base (422: a volume
                measurement's polygon or base), job_running (409: the resource's job is queued
                or running), not_ready (409: the resource has not finished importing),
                range_not_satisfiable (416)
            message: { type: string }
            details:
              type: object
              additionalProperties: true
      example:
        error: { code: not_found, message: "project 7f1c… not found", details: {} }
...
    ClassDefInput:
      type: object
      required: [name, colour]
      properties:
        id: { type: string, description: keep the id to rename or recolour an existing class }
        name: { type: string, minLength: 1, maxLength: 64, pattern: '\S', description: must contain a non-whitespace character }
        colour: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        hotkey:
          type: [string, "null"]
          description: single character, 1 to 9 recommended
          maxLength: 1
      example: { name: excavator, colour: "#f97316", hotkey: "1" }
    ClassDef:
      type: object
      required: [id, name, colour, hotkey, order]
      properties:
        id: { type: string }
        name: { type: string }
        colour: { type: string }
        hotkey: { type: [string, "null"] }
        order: { type: integer }
      example: { id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", hotkey: "1", order: 0 }
    ImportSettings:
      type: object
      description: preparation settings (spec section 5); missing fields take the project defaults
      properties:
        max_side: { type: integer, minimum: 512, maximum: 12000, default: 4000 }
        quality: { type: integer, minimum: 50, maximum: 100, default: 95 }
        dedupe_threshold: { type: integer, minimum: 0, maximum: 32, default: 4, description: max phash Hamming distance for a duplicate }
        group_regex:
          type: string
          default: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)"
          description: the named group `flight` is the group key (flights span cameras); fall back to a 250 m tile, then the source site
      example: { max_side: 4000, quality: 95, dedupe_threshold: 4, group_regex: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)" }
    ProjectKind:
      type: string
      enum: [train, detect]
      description: "`train` labels images, builds datasets and trains models; `detect` runs library models over sources. Set at creation, never changed."
      example: train
    Project:
      type: object
      required: [id, name, folder, kind, classes, preannotation_model_id, import_defaults, schema_version, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        folder: { type: string, description: absolute path of the project folder }
        kind: { $ref: "#/components/schemas/ProjectKind" }
        classes:
          type: array
          items: { $ref: "#/components/schemas/ClassDef" }
        preannotation_model_id: { type: [string, "null"], description: a library model id }
        import_defaults: { $ref: "#/components/schemas/ImportSettings" }
        schema_version: { type: integer }
        created_at: { type: string, format: date-time }
      example:
        id: "7f1c2e3a-1111-4000-8000-000000000001"
        name: Ahmadia
        folder: "E:\\Projects\\Ahmadia"
        kind: train
        classes:
          - { id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", hotkey: "1", order: 0 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000002", name: wheel_loader, colour: "#eab308", hotkey: "2", order: 1 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000003", name: bulldozer, colour: "#22c55e", hotkey: "3", order: 2 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000004", name: dump_truck, colour: "#06b6d4", hotkey: "4", order: 3 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000005", name: crane, colour: "#3b82f6", hotkey: "5", order: 4 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000006", name: concrete_mixer, colour: "#a855f7", hotkey: "6", order: 5 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000007", name: roller, colour: "#ec4899", hotkey: "7", order: 6 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000008", name: backhoe, colour: "#ef4444", hotkey: "8", order: 7 }
        preannotation_model_id: "m0000000-2222-4000-8000-000000000001"
        import_defaults: { max_side: 4000, quality: 95, dedupe_threshold: 4, group_regex: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)" }
        schema_version: 1
        created_at: "2026-09-17T10:00:00Z"
    ProjectCreate:
      type: object
      required: [name, folder, classes, kind]
      properties:
        name: { type: string, minLength: 1 }
        folder: { type: string, description: absolute path; created when missing, must not already hold a project }
        kind: { $ref: "#/components/schemas/ProjectKind" }
        classes:
          type: array
          description: a detection project may start with an empty list
          items: { $ref: "#/components/schemas/ClassDefInput" }
      example:
        name: Ahmadia
        folder: "E:\\Projects\\Ahmadia"
        kind: train
        classes:
          - { name: excavator, colour: "#f97316", hotkey: "1" }
          - { name: dump_truck, colour: "#06b6d4", hotkey: "2" }
    ProjectOpen:
      type: object
      required: [folder]
      properties:
        folder: { type: string }
      example: { folder: "E:\\Projects\\Ahmadia" }
    ProjectUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        preannotation_model_id: { type: [string, "null"], description: a library model id }
        import_defaults: { $ref: "#/components/schemas/ImportSettings" }
      example: { name: Ahmadia, preannotation_model_id: "m0000000-2222-4000-8000-000000000001" }
    ProjectPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Project" }
        next_cursor: { type: [string, "null"] }
    Stats:
      type: object
      description: dataset statistics for a project or one source (spec section 5)
      required: [image_count, labeled_count, unlabeled_count, box_count, pending_review_count, duplicate_count, boxes_per_class, sources, groups, resolution_histogram, capture_time_range, gps_bounds]
      properties:
        image_count: { type: integer }
        labeled_count: { type: integer, description: images with at least one accepted or edited box }
        unlabeled_count: { type: integer }
        box_count: { type: integer, description: accepted and edited boxes }
        pending_review_count: { type: integer, description: unreviewed proposals }
        duplicate_count: { type: integer, description: files skipped as near-duplicates at import }
        boxes_per_class:
          type: array
          items:
            type: object
            required: [class_id, class_name, count]
            properties:
              class_id: { type: string }
              class_name: { type: string }
              count: { type: integer }
        sources:
          type: array
          items:
            type: object
            required: [source_id, site, image_count]
            properties:
              source_id: { type: string }
              site: { type: string }
              image_count: { type: integer }
        groups:
          type: array
          items:
            type: object
            required: [group_key, image_count]
            properties:
              group_key: { type: string }
              image_count: { type: integer }
        resolution_histogram:
          type: array
          items:
            type: object
            required: [width, height, count]
            properties:
              width: { type: integer }
              height: { type: integer }
              count: { type: integer }
        capture_time_range:
          type: [object, "null"]
          required: [min, max]
          properties:
            min: { type: string, format: date-time }
            max: { type: string, format: date-time }
        gps_bounds:
          type: [object, "null"]
          required: [min_lat, min_lon, max_lat, max_lon]
          properties:
            min_lat: { type: number }
            min_lon: { type: number }
            max_lat: { type: number }
            max_lon: { type: number }
      example:
        image_count: 3299
        labeled_count: 30
        unlabeled_count: 3269
        box_count: 112
        pending_review_count: 41
        duplicate_count: 0
        boxes_per_class:
          - { class_id: "c1a2b3c4-0000-4000-8000-000000000001", class_name: excavator, count: 40 }
          - { class_id: "c1a2b3c4-0000-4000-8000-000000000004", class_name: dump_truck, count: 72 }
        sources:
          - { source_id: "50000000-3333-4000-8000-000000000001", site: ahmadia, image_count: 3299 }
        groups:
          - { group_key: "0031", image_count: 697 }
          - { group_key: "0033", image_count: 622 }
        resolution_histogram:
          - { width: 4000, height: 2667, count: 3299 }
        capture_time_range: { min: "2019-04-15T06:35:36Z", max: "2019-04-15T09:12:01Z" }
        gps_bounds: { min_lat: 29.4901, min_lon: 47.7602, max_lat: 29.4988, max_lon: 47.7701 }
...